Enabled logging
I know this doesn’t have anything to do with what this branch is for, but it is relevant in figuring out this problem I was having where the nav just wouldn’t work. I think it was a browser cache thing, but I’ll look into it when I get the chance.