Responsive Desktop
So I’ve done a couple of different things here, I’ve got the desktop view completely responsive now using only a vew custom properties and media queries, and I’ve reorganized my CSS to be more efficient. I might make some more adjustments to the text on desktop view, but it’s not the end of the world if I don’t.